# contracts/openapi/providentia-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Providentia API",
    "version": "1.3.0",
    "summary": "Authoritative implemented Phase 2/4 increment contract",
    "description": "Bounded identity, multi-home tenancy, published global catalog reads, and home-scoped offline synchronization prototype."
  },
...
  "servers": [{"url": "/"}],
  "security": [{"bearerAuth": []}, {"cookieAuth": []}],
  "paths": {
...
    "/api/v1/auth/verify-email": {
      "post": {
        "operationId": "verifyEmail",
        "security": [],
        "tags": ["Identity"],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TokenRequest"}}}},
        "responses": {"204": {"description": "Verified"}, "422": {"$ref": "#/components/responses/Problem"}}
      }
    },
...
    "/api/v1/auth/password-reset/request": {
      "post": {
        "operationId": "requestPasswordReset",
        "security": [],
        "tags": ["Identity"],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/EmailRequest"}}}},
        "responses": {"202": {"description": "Accepted without account enumeration"}}
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {"type": "http", "scheme": "bearer", "bearerFormat": "opaque"},
      "cookieAuth": {"type": "apiKey", "in": "cookie", "name": "providentia_access"}
    },
...
    "responses": {
      "Problem": {"description": "RFC 9457 Problem Details", "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/ProblemDetails"}}}}
    },
    "schemas": {
      "ProblemDetails": {"type": "object", "description": "Problem Details for HTTP APIs as defined by RFC 9457.", "required": ["type", "title", "status", "requestId"], "properties": {"type": {"type": "string"}, "title": {"type": "string"}, "status": {"type": "integer", "minimum": 400, "maximum": 599}, "detail": {"type": "string"}, "requestId": {"type": "string"}}, "additionalProperties": true},
...
      "EmailRequest": {"type": "object", "additionalProperties": false, "required": ["email"], "properties": {"email": {"type": "string", "format": "email"}}},
      "TokenRequest": {"type": "object", "additionalProperties": false, "required": ["token"], "properties": {"token": {"type": "string", "minLength": 40}}},
...
    }
  },
...
}